+++ wemake_python_styleguide/violations/best_practices.py
@@ -80,6 +80,7 @@ Summary
    ForbiddenInlineIgnoreViolation
    WrongMultilineStringUseViolation
    GetterWithoutReturnViolation
+   EmptyCommentViolation
 
 Best practices
 --------------
@@ -148,6 +149,7 @@ Best practices
 .. autoclass:: ForbiddenInlineIgnoreViolation
 .. autoclass:: WrongMultilineStringUseViolation
 .. autoclass:: GetterWithoutReturnViolation
+.. autoclass:: EmptyCommentViolation
 
 """
 
@@ -2465,3 +2467,40 @@ class GetterWithoutReturnViolation(ASTViolation):
 
     error_template = 'Found a getter without a return value'
     code = 463
+
+
+@final
+class EmptyCommentViolation(TokenizeViolation):
+    """
+    Forbid empty comments.
+
+    Empty comments are only allowed in between valid comments.
+
+    Reasoning:
+        Empty comments that do not help formatting should be excluded.
+
+    Solution:
+        Remove the empty comments.
+
+    Example::
+
+        # Correct:
+
+        # First line
+        #
+        # Samples:
+        # One
+        # Two
+        my_var = 1
+
+        # Wrong:
+
+        #
+        my_var = 1
+
+    .. versionadded:: 0.15.0
+
+    """
+
+    error_template = 'Found empty comment'
+    code = 464